# docs/swagger.yaml
basePath: /api
definitions:
  models.APIKey:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      key:
        description: Only show on creation
        type: string
      last_used:
        type: string
      name:
        type: string
      profile_id:
        type: string
      subscription_plan:
        description: Added for rate limiting context
        type: string
      updated_at:
        type: string
    type: object
...
  models.CreateAPIKeyRequest:
    properties:
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - name
    type: object
  models.CreateAPIKeyResponse:
    properties:
      id:
        type: string
      key:
        type: string
      message:
        type: string
      name:
        type: string
    type: object
...
  models.PaymentHistoryResponse:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      payments:
        items:
          $ref: '#/definitions/models.Payment'
        type: array
      total_count:
        type: integer
    type: object
...
host: 146.190.186.116:8000
info:
  contact:
    email: ifalabstudio@gmail.com
    name: IfaLabs
    url: https://ifalabs.com
  description: IFA LABS Oracle Engine API provides real-time, reliable asset prices
    using an aggregated moving window algorithm to ensure stability and reduce manipulation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Oracle Engine API
  version: "1.0"
paths:
...
  /dashboard/{id}/api-keys:
    get:
      description: Get all API keys for a profile
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.APIKey'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get API Keys
      tags:
      - dashboard
    post:
      consumes:
      - application/json
      description: Create a new API key for accessing the Oracle Engine API
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      - description: Create API key request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateAPIKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CreateAPIKeyResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create API Key
      tags:
      - dashboard
  /dashboard/{id}/api-keys/{key_id}:
    delete:
      description: Delete an API key
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      - description: API Key ID
        in: path
        name: key_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete API Key
      tags:
      - dashboard
  /dashboard/{id}/payment:
    post:
      consumes:
      - application/json
      description: Create a new payment (placeholder implementation)
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      - description: Create payment request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreatePaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Payment'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create Payment
      tags:
      - dashboard
  /dashboard/{id}/payment/history:
    get:
      description: Get payment history for a profile
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 20, max: 100)'
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaymentHistoryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get Payment History
      tags:
      - dashboard
  /dashboard/{id}/profile:
    get:
      description: Get company profile information
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CompanyProfile'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get User Profile
      tags:
      - dashboard
    put:
      consumes:
      - application/json
      description: Update company profile information
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      - description: Update profile request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CompanyProfile'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update User Profile
      tags:
      - dashboard
  /dashboard/{id}/usage:
    get:
      consumes:
      - application/json
      description: Returns API usage statistics for the authenticated user
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get API usage statistics
      tags:
      - dashboard
  /dashboard/login:
    post:
      consumes:
      - application/json
      description: Login with email and password
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User Login
      tags:
      - dashboard
  /dashboard/signup:
    post:
      consumes:
      - application/json
      description: Create a new company profile and user account
      parameters:
      - description: Sign up request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SignUpResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User Sign Up a company
      tags:
      - dashboard
  /issuances/{id}:
    get:
      consumes:
      - application/json
      description: Returns details of a specific issuance
      parameters:
      - description: Issuance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Issuance'
      summary: Get issuance details
      tags:
      - issuances
  /prices/{id}/audit:
    get:
      consumes:
      - application/json
      description: Returns audit information for a specific price
      parameters:
      - description: Price ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PriceAudit'
      summary: Get price audit
      tags:
      - prices
  /prices/last:
    get:
      consumes:
      - application/json
      description: Returns the last known price for a specific asset or all assets
      parameters:
      - description: Asset ID to get price for
        in: query
        name: asset
        type: string
      - description: Comma-separated list of price change periods (e.g. '7d,3d,24h').
          Default is '7d'
        in: query
        name: changes
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/models.UnifiedPrice'
            type: object
      summary: Get last price for an asset
      tags:
      - prices
...
securityDefinitions:
  ApiKeyAuth:
    description: API key for accessing Oracle Engine endpoints.
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"